Create a separate client struct
To keep the connection to FCM alive and for a better throughput it is important
to hold the connection in a separate struct. This change creates a new struct
Client, which holds the HTTP connection and has a public interface `send` for
sending messages to the FCM service.
