+++ src/message/tests.rs
@@ -1,8 +1,6 @@
 use {Message, Priority};
-use FcmError;
 use notification::NotificationBuilder;
 use std::collections::HashMap;
-use hyper::status::StatusCode;
 
 #[test]
 fn should_create_new_message() {
@@ -136,54 +134,3 @@ fn should_set_notifications() {
     assert!(msg.notification != None);
 }
 
-#[test]
-fn should_parse_error_as_unauthorized() {
-    let result = Message::parse_response(StatusCode::Unauthorized, "Unauthorized");
-
-    assert!(result.is_err());
-    assert_eq!(result.err().unwrap(), FcmError::Unauthorized);
-}
-
-#[test]
-fn should_parse_error_as_invalid_message() {
-    let result = Message::parse_response(StatusCode::BadRequest, "INVALID_REGISTRATION");
-
-    assert!(result.is_err());
-    assert_eq!(result.err().unwrap(), 
-    FcmError::InvalidMessage("INVALID_REGISTRATION".to_string()));
-}
-
-#[test]
-fn should_parse_error_as_server_error() {
-    let result = Message::parse_response(StatusCode::InternalServerError, "Internal Server Error");
-
-    assert!(result.is_err());
-    assert_eq!(result.err().unwrap(), FcmError::ServerError);
-}
-
-#[test]
-fn should_parse_successful_response() {
-    let response = r#"
-        {
-            "message_id": 2000000,
-            "results": [
-                {
-                    "message_id": 200000,
-                    "registration_id": 200000,
-                    "error": "error"
-                }
-            ]
-        }
-    "#;
-    let result = Message::parse_response(StatusCode::Ok, response);
-
-    assert!(result.is_ok());
-
-    let result = result.unwrap();
-
-    assert_eq!(result.message_id.unwrap(), 2000000);
-
-    let message_results = result.results.unwrap();
-
-    assert_eq!(message_results.len(), 1);
-}
